feat: 修改 DBPortMapper 异常处理问题; DBListenPort API 迁移至 terminal app 中
