+++ apps/applications/urls/api_urls.py
@@ -13,7 +13,6 @@ router.register(r'applications', api.ApplicationViewSet, 'application')
 router.register(r'accounts', api.ApplicationAccountViewSet, 'application-account')
 router.register(r'system-users-apps-relations', api.SystemUserAppRelationViewSet, 'system-users-apps-relation')
 router.register(r'account-secrets', api.ApplicationAccountSecretViewSet, 'application-account-secret')
-router.register(r'db-listen-ports', api.DBListenPortViewSet, 'db-listen-ports')
 
 
 urlpatterns = [
