+++ apps/terminal/api/__init__.py
@@ -8,3 +8,4 @@ from .storage import *
 from .status import *
 from .sharing import *
 from .endpoint import *
+from .db_listen_port import *
